Mapper to Converter

+++ product-search-service/src/main/java/com/konkon/onlinestore/product/search/service/application/rest/model/CategoryResponse.java
@@ -1,6 +1,6 @@
 package com.konkon.onlinestore.product.search.service.application.rest.model;
 
 public record CategoryResponse(
-        Long categoryId,
+        int categoryId,
         String categoryName) {
 }
